requirements.txt
fastapi
uvicorn
ruff
mysql-connector-python